remove share file from menu on shared file tab
