+++ client/src/types/FilePath.ts
@@ -1,0 +1,6 @@
+enum FilePath {
+	Owned = 'ownedFiles',
+	Shared = 'sharedFiles'
+}
+
+export default FilePath;
