+++ client/src/types/RouteParams.ts
@@ -1,0 +1,7 @@
+import FilePath from "./FilePath";
+
+interface RouteParams {
+	ownedOrShared: FilePath;
+}
+
+export default RouteParams
